requirements.txt
alembic==1.8.1
anyio==3.6.1
autopep8==1.6.0
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.85.1
fastapi-pagination==0.10.0
h11==0.13.0
idna==3.3
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2==2.9.3
pycodestyle==2.9.0
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
python-decouple==3.6
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.20.4
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.2
